add simple energy parsers
